requirements.txt
memory-profiler==0.61.0
pytest==8.2.2
pylint==3.2.5
mypy==1.10.1
isort==5.13.2
black==24.4.2
pandas==2.2.2
google-cloud-bigquery==3.25.0
